refactoring plotly display
